+++ consumer/src/main/java/at/uibk/dps/consumer/module/policy/update/dto/UpdatePolicyRequestDto.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class UpdatePolicyRequestDto {
-    private long fetchMinBytes;
+    private int fetchMinBytes;
     private int fetchMaxWaitMs;
-    private long maxPollRecords;
+    private int maxPollRecords;
     private int maxPollIntervalMs;
     private int sessionTimeoutMs;
 }
